perf(route): refactor guard
