+++ src/components/Modal/src/hooks/useModalFullScreen.ts
@@ -7,7 +7,7 @@ export interface UseFullScreenContext {
 }
 
 export function useFullScreen(context: UseFullScreenContext) {
-  const formerHeightRef = ref(0);
+  // const formerHeightRef = ref(0);
   const fullScreenRef = ref(false);
 
   const getWrapClassName = computed(() => {
@@ -20,25 +20,25 @@ export function useFullScreen(context: UseFullScreenContext) {
     e && e.stopPropagation();
     fullScreenRef.value = !unref(fullScreenRef);
 
-    const modalWrapper = unref(context.modalWrapperRef);
+    // const modalWrapper = unref(context.modalWrapperRef);
 
-    if (!modalWrapper) return;
+    // if (!modalWrapper) return;
 
-    const wrapperEl = modalWrapper.$el as HTMLElement;
-    if (!wrapperEl) return;
-    const modalWrapSpinEl = wrapperEl.querySelector('.ant-spin-nested-loading') as HTMLElement;
+    // const wrapperEl = modalWrapper.$el as HTMLElement;
+    // if (!wrapperEl) return;
+    // const modalWrapSpinEl = wrapperEl.querySelector('.ant-spin-nested-loading') as HTMLElement;
 
-    if (!modalWrapSpinEl) return;
+    // if (!modalWrapSpinEl) return;
 
-    if (!unref(formerHeightRef) && unref(fullScreenRef)) {
-      formerHeightRef.value = modalWrapSpinEl.offsetHeight;
-    }
+    // if (!unref(formerHeightRef) && unref(fullScreenRef)) {
+    //   formerHeightRef.value = modalWrapSpinEl.offsetHeight;
+    // }
 
-    if (unref(fullScreenRef)) {
-      modalWrapSpinEl.style.height = `${window.innerHeight - unref(context.extHeightRef)}px`;
-    } else {
-      modalWrapSpinEl.style.height = `${unref(formerHeightRef)}px`;
-    }
+    // if (unref(fullScreenRef)) {
+    //   modalWrapSpinEl.style.height = `${window.innerHeight - unref(context.extHeightRef)}px`;
+    // } else {
+    //   modalWrapSpinEl.style.height = `${unref(formerHeightRef)}px`;
+    // }
   }
   return { getWrapClassName, handleFullScreen, fullScreenRef };
 }
