+++ src/views/demo/comp/modal/Modal1.vue
@@ -8,8 +8,5 @@
   import { BasicModal } from '/@/components/Modal';
   export default defineComponent({
     components: { BasicModal },
-    setup() {
-      return {};
-    },
   });
 </script>
